src/job/find_it/requirements.txt
beautifulsoup4==4.6.0
dj-database-url==0.5.0
Django==2.2.14
django-jsonfield==1.4.0
djangorestframework
elasticsearch==6.3.1
gunicorn==19.8.1
lazy-object-proxy==1.3.1
psycopg2-binary
requests==2.20.0
sqlparse==0.3.0
urllib3==1.24.2
whitenoise==4.1.2
